requirements.txt
streamlit
# LLM & API
openai-whisper
sentence-transformers

# File Loaders
pypdf

# Databases
chromadb
psycopg2-binary

# Frontend
streamlit

# Utilities
python-dotenv
reportlab
langchain-chroma

openai-whisper
ffmpeg-python==0.2.0
google-generativeai
langgraph
langchain
google-generativeai
langchain-google-genai
langchain-community
networkx
pyvis
langchain-core
spacy